The refactoring in the registration is done.
