+++ src/views/Register/components/AddressSection/StateField.tsx
@@ -1,0 +1,52 @@
+import Grid from "@mui/material/Grid";
+import TextField from "@mui/material/TextField";
+import { CustomLabel } from "@/components";
+const StateField = ({ formik }) => {
+  const primaryColor = "#009999";
+  const placeholderColor = "rgba(51, 51, 51, 0.4)";
+
+  const customTextField = {
+    width: "90%",
+    "& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+      borderRadius: 0,
+      border: "2px solid " + primaryColor,
+    },
+    "&:hover .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+    },
+    "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+    },
+    "& .MuiInputLabel-outlined": {
+      fontSize: "1rem",
+      color: placeholderColor,
+    },
+    "& .MuiInputLabel-outlined.Mui-focused": {
+      color: primaryColor,
+    },
+    "& .MuiOutlinedInput-input": {
+      padding: "0.7rem",
+    },
+  };
+
+  return (
+    <Grid item xs={12} sm={6} md={4}>
+      <CustomLabel name="State" required={true} />
+      <TextField
+        type="text"
+        onChange={formik.handleChange}
+        placeholder="State"
+        name="addressState"
+        value={formik.values.addressState}
+        sx={customTextField}
+        error={
+          formik.touched.addressState && Boolean(formik.errors.addressState)
+        }
+        helperText={formik.touched.addressState && formik.errors.addressState}
+      />
+    </Grid>
+  );
+};
+
+export default StateField;
