+++ src/views/Register/components/SecondLastName.tsx
@@ -1,0 +1,54 @@
+import Grid from "@mui/material/Grid";
+import TextField from "@mui/material/TextField";
+import { CustomLabel } from "@/components";
+const SecondLastNameField = ({ formik }) => {
+  const primaryColor = "#009999";
+  const placeholderColor = "rgba(51, 51, 51, 0.4)";
+
+  const customTextField = {
+    width: "90%",
+    "& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+      borderRadius: 0,
+      border: "2px solid " + primaryColor,
+    },
+    "&:hover .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+    },
+    "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline": {
+      borderColor: primaryColor,
+    },
+    "& .MuiInputLabel-outlined": {
+      fontSize: "1rem",
+      color: placeholderColor,
+    },
+    "& .MuiInputLabel-outlined.Mui-focused": {
+      color: primaryColor,
+    },
+    "& .MuiOutlinedInput-input": {
+      padding: "0.7rem",
+    },
+  };
+
+  return (
+    <Grid item xs={12} sm={6} md={4}>
+      <CustomLabel name="Second Last Name" required={false} />
+      <TextField
+        sx={customTextField}
+        name="secondLastName"
+        type="text"
+        placeholder="Second Last Name"
+        onChange={formik.handleChange}
+        value={formik.values.secondLastName}
+        error={
+          formik.touched.secondLastName && Boolean(formik.errors.secondLastName)
+        }
+        helperText={
+          formik.touched.secondLastName && formik.errors.secondLastName
+        }
+      />
+    </Grid>
+  );
+};
+
+export default SecondLastNameField;
